+++ src/app/mock-user.ts
@@ -1,9 +1,10 @@
-export class User {
-  id: number;
-  name: string;
-  lastname: string;
-  profileimg: string;
-  email: string;
-  birthday: date;
-  password: string;
-}
+// Usa la clase User por supuesto
+import { User } from './user';
+// Genera una constante  del objeto User
+export const USER: User[] = [
+  { id: 1, name: 'William', lastname: 'González' , profileimg: '../assets/imgs/mifoto.jpg', email: "willigon106@gmail.com", birthday: '25-06-1995', password: 'cosito'}
+  /*{ id: 2, name: 'Huerta 2', description: default_description , image: 'nota-2-huertas.jpg', icon: "leaf"},
+  { id: 3, name: 'Huerta 3', description: default_description , image: 'nota-3-huertas.jpg', icon: "leaf"},
+  { id: 4, name: 'Huerta 4', description: default_description , image: 'nota-4-huertas.jpg', icon: "leaf"},
+  { id: 5, name: 'Huerta 5', description: default_description , image: 'nota-5-huertas.jpg', icon: "leaf"}*/
+];
