+++ src/app/user.ts
@@ -1,0 +1,9 @@
+export class User {
+  id: number;
+  name: string;
+  lastname: string;
+  profileimg: string;
+  email: string;
+  birthday: date;
+  password: string;
+}
