+++ src/pages/profile/profile.html
@@ -15,16 +15,13 @@
 <ion-content>
   <ion-list>
 	<ion-item>
-        <!--<ion-avatar item-start>
-          <img src="{{character.image}}">
+        <ion-avatar item-start>
+          <img src="{{User.profileimg}}">
         </ion-avatar>
-		
-        <h2>{{characters.name}}</h2>
-		<h2>{{characters.lastname}}</h2>
-        <p>{{characters.email}}</p>
-		<p>{{characters.birthday}}</p>
-		<p>{{characters.password}}</p>
-		-->
+        <h2>{{User.name}}</h2>
+		<h2>{{User.lastname}}</h2>
+        <p>{{User.email}}</p>
+		<p>{{User.birthday}}</p>
     </ion-item>
   </ion-list>
 </ion-content>
